+++ src/lib/fauna/client.js
@@ -1,0 +1,9 @@
+require('dotenv').config();
+const { Client } = require('fauna');
+
+// Create a FaunaDB client instance
+const client = new Client({
+  secret: process.env.FAUNA_SECRET_KEY,
+});
+
+module.exports = client;